Fix rotation error
Problem on some devices with old hack (locking rotation during clean so activity
won't be regenerated on screen rotation). Replacing it with a slightly less ugly
hack (simply telling Android to not regenerate activity on rotation). This will
have to replaced in a future re-write as it is still a band-aid and not a true
solution